requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
anyio==3.3.1
mysql-connector
appdirs==1.4.4
async-exit-stack==1.0.1
async-generator==1.10
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
black==21.7b0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
cryptography==3.4.8
dataclasses==0.6
dnspython==2.1.0
email-validator==1.1.3
et-xmlfile==1.1.0
fastapi==0.68.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.1
h11==0.12.0
httpcore==0.13.6
httptools==0.1.2
httpx==0.19.0
idna==3.2
importlib-metadata==4.6.3
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
lxml==4.*
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.26
numpy==1.21.2
openpyxl==3.0.7
orjson==3.6.1
pandas
paramiko==2.7.2
pathspec==0.9.0
playwright==1.14.1
promise==2.3
protobuf==3.17.3
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.8.2
pyee==8.2.2
PyJWT==2.1.0
PyNaCl
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.3
requests==2.26.0
rfc3986==1.5.0
Rx==1.6.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
starlette==0.14.2
tomli==1.2.0
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.6
uvicorn==0.13.4
uvloop==0.15.3
uWSGI==2.0.19.1
watchgod==0.7
websockets==8.1
Werkzeug==2.0.1
xlrd==2.0.1
XlsxWriter==3.0.2
xmltodict==0.12.0
zipp==3.5.0
